cli/requirements.txt
click
python-frontmatter
GitPython
